requirements_production.txt
# AMIEN Production Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-run==0.10.0
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.16.4
google-cloud-scheduler==2.13.4
google-cloud-monitoring==2.16.0
google-generativeai==0.3.2
openai==1.3.7
requests==2.31.0
aiohttp==3.9.1
asyncio==3.4.3
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
scipy==1.11.4
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.1
transformers==4.36.2
datasets==2.14.6
accelerate==0.25.0
